agents/requirements.txt
pydantic-ai==0.0.14
pydantic==2.9.0
fastapi==0.115.0
uvicorn==0.30.0
httpx==0.27.0
python-dotenv==1.0.1
solders==0.21.0
base58==2.1.1
requests==2.32.0
